+++ changelog.md
@@ -2,6 +2,13 @@
 
 All notable changes to this mod will be documented in this file.
 
+## [0.1.21] - 2026-05-06
+
+- Optimization pass
+  - Commodity names cached as a static once-generated map on game load
+  - Reserved commodity mission intel scanning cached per-frame
+  - Fleet data cached per-frame
+
 ## [0.1.20] - 2026-05-05
 
 - Ability tooltip rework concluded